requirements.txt
pytorch
transformers
protobuf
accelerate
sentencepiece
